compute/requirements.txt
swan-sdk
python-dotenv
pytest